+++ src/client/index.ts
@@ -1,5 +1,6 @@
 import './devtools.js';
 import App from './App.svelte';
+import './test_imports/test_imports.js';
 
 const root = document.getElementById('root');
 if (!root) throw Error('Cannot find root element');